Split correlation into runner script and functions library
